wip: add mining content |m move files
