--- src/containers/Airdrop/AirdropPermissionSettings/styles.ts
@@ -1,41 +1,0 @@
-import styled from 'styled-components';
-
-export const Wrapper = styled('div')`
-    padding: 0 0 0px 0;
-`;
-
-export const BoxWrapper = styled('div')`
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    gap: 15px;
-
-    padding: 20px;
-
-    border-radius: 10px;
-    background: #fff;
-    box-shadow: 0px 0px 25px 0px rgba(0, 0, 0, 0.05);
-    width: 100%;
-
-    margin-bottom: 20px;
-`;
-
-export const TextWrapper = styled('div')`
-    display: flex;
-    flex-direction: column;
-    gap: 4px;
-`;
-
-export const Title = styled('div')`
-    color: #000;
-    font-size: 14px;
-    font-weight: 600;
-    line-height: 110%;
-`;
-
-export const Text = styled('span')`
-    color: #797979;
-    font-size: 12px;
-    font-weight: 500;
-    line-height: 116.667%;
-`;
