+++ src/containers/Settings/GeneralSettings.tsx
@@ -1,7 +1,7 @@
-import AirdropPermissionSettings from '@app/containers/Airdrop/AirdropPermissionSettings/AirdropPermissionSettings.tsx';
-import LogsSettings from '@app/containers/SideBar/components/Settings/LogsSettings.tsx';
-import LanguageSettings from '@app/containers/SideBar/components/Settings/LanguageSettings.tsx';
-import { ResetSettingsButton } from '@app/containers/SideBar/components/Settings/ResetSettingsButton.tsx';
+import AirdropPermissionSettings from './sections/general/AirdropPermissionSettings.tsx';
+import LogsSettings from './sections/general/LogsSettings.tsx';
+import LanguageSettings from './sections/general/LanguageSettings.tsx';
+import { ResetSettingsButton } from './sections/general/ResetSettingsButton.tsx';
 
 export const GeneralSettings = () => {
     return (
