+++ src/containers/Settings/sections/mining/GpuMiningMarkup.tsx
